comment, refactor, cleanup
