+++ modules/offers/client/views/directives/tr-offer-host-view.client.view.html
@@ -188,10 +188,7 @@
 
       </div>
 
-      <!-- The map -->
-      <div tr-offer-location="trOfferHost.offer"
-        ng-if="trOfferHost.offer.status === 'yes' || trOfferHost.offer.status === 'maybe'"></div>
-
+      <!-- The map (React component) -->
       <offer-location
         ng-if="trOfferHost.offer.status === 'yes' || trOfferHost.offer.status === 'maybe'"
         offer="trOfferHost.offer"
