+++ modules/offers/client/views/offer-meet-list.client.view.html
@@ -43,8 +43,8 @@
        ng-attr-id="offer-{{::offer._id}}">
     <div class="panel-body">
 
-      <!-- The map -->
-      <div tr-offer-location="offer"></div>
+      <!-- The map (React component) -->
+      <offer-location offer="offer"></offer-location>
 
       <br><br>
